made changes to show history from drawer
