+++ lib/components/background_painter.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class BackgroundPainter extends CustomPainter {
+  @override
+  void paint(Canvas canvas, Size size) {
+    final paint = Paint()
+      ..color = Colors.blue
+      ..style = PaintingStyle.fill;
+
+    final path = Path()
+      ..moveTo(0, 50)
+      ..quadraticBezierTo(size.width / 2, size.height * 0.4, size.width, 50)
+      ..lineTo(size.width, 0)
+      ..lineTo(0, 0)
+      ..close();
+
+    canvas.drawPath(path, paint);
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) {
+    return false;
+  }
+}
