+++ lib/screens/dashboard.dart
@@ -1,6 +1,7 @@
 import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:attendance_app/screens/drawer.dart';
+import 'package:attendance_app/components/background_painter.dart';
 
 class Dashboard extends StatefulWidget {
   const Dashboard({super.key});
@@ -126,29 +127,6 @@ class ProfileCard extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-class BackgroundPainter extends CustomPainter {
-  @override
-  void paint(Canvas canvas, Size size) {
-    final paint = Paint()
-      ..color = Colors.blue
-      ..style = PaintingStyle.fill;
-
-    final path = Path()
-      ..moveTo(0, 50)
-      ..quadraticBezierTo(size.width / 2, size.height * 0.4, size.width, 50)
-      ..lineTo(size.width, 0)
-      ..lineTo(0, 0)
-      ..close();
-
-    canvas.drawPath(path, paint);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return false;
   }
 }
